requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
drf-spectacular==0.21.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
jsonschema==4.2.1
kombu==5.2.2
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
pika==1.2.0
prompt-toolkit==3.0.24
psycopg2==2.9.2
pymongo==4.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
